# Move Rosbag IO into background thread in offline_node_main.