--- a/cartographer_ros/cartographer_ros/offline_node_main.cc
+++ b/cartographer_ros/cartographer_ros/offline_node_main.cc
@@ -24,6 +24,7 @@
 #include <string>
 #include <vector>
 
+#include "cartographer/common/blocking_queue.h"
 #include "cartographer/common/make_unique.h"
 #include "cartographer/common/port.h"
 #include "cartographer_ros/node.h"
@@ -68,7 +69,195 @@ constexpr char kTfTopic[] = "tf";
 constexpr double kClockPublishFrequencySec = 1. / 30.;
 constexpr int kSingleThreaded = 1;
 
-void Run(const std::vector<std::string>& bag_filenames) {
+// A poor mans typed union that signifies events from the IO thread.
+struct Event {
+  enum class Kind {
+    kNewBag,
+    kBagEnd,
+    kRosMessage,
+  };
+
+  // A poors mans union that contains instantiated ROS messages.
+  struct RosMessage {
+    enum class Kind {
+      kTf,
+      kImu,
+      kOdometry,
+      kLaserScan,
+      kMultiEchoLaserScan,
+      kPointCloud2,
+    };
+
+    RosMessage::Kind kind;
+    string topic;
+    ros::Time serialization_time;
+    sensor_msgs::Imu::ConstPtr imu;
+    sensor_msgs::LaserScan::ConstPtr laser_scan;
+    sensor_msgs::MultiEchoLaserScan::ConstPtr multi_echo_laser_scan;
+    sensor_msgs::PointCloud2::ConstPtr point_cloud2;
+    nav_msgs::Odometry::ConstPtr odometry;
+    tf2_msgs::TFMessage::ConstPtr tf;
+  };
+
+  static Event NewBag(::ros::Time begin_time, double duration_in_seconds) {
+    Event event;
+    event.kind = Kind::kNewBag;
+    event.new_bag.begin_time = begin_time;
+    event.new_bag.duration_in_seconds = duration_in_seconds;
+    return event;
+  }
+
+  static Event BagEnd() {
+    Event event;
+    event.kind = Kind::kBagEnd;
+    return event;
+  }
+
+  static Event TfMessage(const string& topic,
+                         const ros::Time serialization_time,
+                         tf2_msgs::TFMessage::ConstPtr ros_message) {
+    CHECK(ros_message != nullptr);
+    Event event;
+    event.kind = Kind::kRosMessage;
+    event.message.kind = RosMessage::Kind::kTf;
+    event.message.topic = topic;
+    event.message.serialization_time = serialization_time;
+    event.message.tf = std::move(ros_message);
+    return event;
+  }
+
+  static Event LaserScanMessage(const string& topic,
+                                const ros::Time serialization_time,
+                                sensor_msgs::LaserScan::ConstPtr ros_message) {
+    CHECK(ros_message != nullptr);
+    Event event;
+    event.kind = Kind::kRosMessage;
+    event.message.kind = RosMessage::Kind::kLaserScan;
+    event.message.topic = topic;
+    event.message.serialization_time = serialization_time;
+    event.message.laser_scan = std::move(ros_message);
+    return event;
+  }
+
+  static Event MultiEchoLaserScanMessage(
+      const string& topic, const ros::Time serialization_time,
+      sensor_msgs::MultiEchoLaserScan::ConstPtr ros_message) {
+    CHECK(ros_message != nullptr);
+    Event event;
+    event.kind = Kind::kRosMessage;
+    event.message.kind = RosMessage::Kind::kMultiEchoLaserScan;
+    event.message.topic = topic;
+    event.message.serialization_time = serialization_time;
+    event.message.multi_echo_laser_scan = std::move(ros_message);
+    return event;
+  }
+
+  static Event PointCloud2Message(
+      const string& topic, const ros::Time serialization_time,
+      sensor_msgs::PointCloud2::ConstPtr ros_message) {
+    CHECK(ros_message != nullptr);
+    Event event;
+    event.kind = Kind::kRosMessage;
+    event.message.kind = RosMessage::Kind::kPointCloud2;
+    event.message.topic = topic;
+    event.message.serialization_time = serialization_time;
+    event.message.point_cloud2 = std::move(ros_message);
+    return event;
+  }
+
+  static Event ImuMessage(const string& topic,
+                          const ros::Time serialization_time,
+                          sensor_msgs::Imu::ConstPtr ros_message) {
+    CHECK(ros_message != nullptr);
+    Event event;
+    event.kind = Kind::kRosMessage;
+    event.message.kind = RosMessage::Kind::kImu;
+    event.message.topic = topic;
+    event.message.serialization_time = serialization_time;
+    event.message.imu = std::move(ros_message);
+    return event;
+  }
+
+  static Event OdometryMessage(const string& topic,
+                               const ros::Time serialization_time,
+                               nav_msgs::Odometry::ConstPtr ros_message) {
+    CHECK(ros_message != nullptr);
+    Event event;
+    event.kind = Kind::kRosMessage;
+    event.message.kind = RosMessage::Kind::kOdometry;
+    event.message.topic = topic;
+    event.message.serialization_time = serialization_time;
+    event.message.odometry = std::move(ros_message);
+    return event;
+  }
+
+  Kind kind;
+
+  struct {
+    ::ros::Time begin_time;
+    double duration_in_seconds;
+  } new_bag;
+  RosMessage message;
+};
+
+void RosbagIoThread(const std::vector<string>& bag_filenames,
+                    const std::unordered_set<string>& expected_sensor_ids,
+                    ::ros::NodeHandle* node_handle,
+                    ::cartographer::common::BlockingQueue<Event>* queue) {
+  for (size_t bag_index = 0; bag_index < bag_filenames.size(); ++bag_index) {
+    const string& bag_filename = bag_filenames[bag_index];
+    if (!::ros::ok()) {
+      break;
+    }
+
+    rosbag::Bag bag;
+    bag.open(bag_filename, rosbag::bagmode::Read);
+    rosbag::View view(bag);
+    const ::ros::Time begin_time = view.getBeginTime();
+    queue->Push(
+        Event::NewBag(begin_time, (view.getEndTime() - begin_time).toSec()));
+
+    for (const rosbag::MessageInstance& msg : view) {
+      if (!::ros::ok()) {
+        break;
+      }
+      const string topic =
+          node_handle->resolveName(msg.getTopic(), false /* resolve */);
+
+      if (msg.isType<tf2_msgs::TFMessage>()) {
+        queue->Push(Event::TfMessage(topic, msg.getTime(),
+                                     msg.instantiate<tf2_msgs::TFMessage>()));
+        continue;
+      }
+
+      if (expected_sensor_ids.count(topic) == 0) {
+        continue;
+      }
+
+      if (msg.isType<sensor_msgs::LaserScan>()) {
+        queue->Push(Event::LaserScanMessage(
+            topic, msg.getTime(), msg.instantiate<sensor_msgs::LaserScan>()));
+      } else if (msg.isType<sensor_msgs::MultiEchoLaserScan>()) {
+        queue->Push(Event::MultiEchoLaserScanMessage(
+            topic, msg.getTime(),
+            msg.instantiate<sensor_msgs::MultiEchoLaserScan>()));
+      } else if (msg.isType<sensor_msgs::PointCloud2>()) {
+        queue->Push(Event::PointCloud2Message(
+            topic, msg.getTime(), msg.instantiate<sensor_msgs::PointCloud2>()));
+      } else if (msg.isType<sensor_msgs::Imu>()) {
+        queue->Push(Event::ImuMessage(topic, msg.getTime(),
+                                      msg.instantiate<sensor_msgs::Imu>()));
+      } else if (msg.isType<nav_msgs::Odometry>()) {
+        queue->Push(Event::OdometryMessage(
+            topic, msg.getTime(), msg.instantiate<nav_msgs::Odometry>()));
+      }
+    }
+    queue->Push(Event::BagEnd());
+    bag.close();
+  }
+}
+
+void Run(const std::vector<string>& bag_filenames) {
   const std::chrono::time_point<std::chrono::steady_clock> start_time =
       std::chrono::steady_clock::now();
   NodeOptions node_options;
@@ -104,6 +293,11 @@ void Run(const std::vector<std::string>& bag_filenames) {
               .second);
   }
 
+  ::cartographer::common::BlockingQueue<Event> queue(100);
+  std::thread rosbag_io_thread(RosbagIoThread, std::ref(bag_filenames),
+                               std::ref(expected_sensor_ids),
+                               node.node_handle(), &queue);
+
   ::ros::Publisher tf_publisher =
       node.node_handle()->advertise<tf2_msgs::TFMessage>(
           kTfTopic, kLatestOnlyPublisherQueueSize);
@@ -128,97 +322,105 @@ void Run(const std::vector<std::string>& bag_filenames) {
       },
       false /* oneshot */, false /* autostart */);
 
-  for (const std::string& bag_filename : bag_filenames) {
-    if (!::ros::ok()) {
-      break;
-    }
+  ::ros::Time begin_time;
+  double duration_in_seconds;
+  int trajectory_id = -1;
 
-    const int trajectory_id =
-        node.AddOfflineTrajectory(expected_sensor_ids, trajectory_options);
+  // We need to keep 'tf_buffer' small because it becomes very inefficient
+  // otherwise. We make sure that tf_messages are published before any data
+  // messages, so that tf lookups always work.
+  std::deque<Event::RosMessage> delayed_messages;
 
-    rosbag::Bag bag;
-    bag.open(bag_filename, rosbag::bagmode::Read);
-    rosbag::View view(bag);
-    const ::ros::Time begin_time = view.getBeginTime();
-    const double duration_in_seconds = (view.getEndTime() - begin_time).toSec();
+  // We publish tf messages one second earlier than other messages. Under
+  // the assumption of higher frequency tf this should ensure that tf can
+  // always interpolate.
+  const ::ros::Duration kDelay(1.);
 
-    // We need to keep 'tf_buffer' small because it becomes very inefficient
-    // otherwise. We make sure that tf_messages are published before any data
-    // messages, so that tf lookups always work.
-    std::deque<rosbag::MessageInstance> delayed_messages;
-    // We publish tf messages one second earlier than other messages. Under
-    // the assumption of higher frequency tf this should ensure that tf can
-    // always interpolate.
-    const ::ros::Duration kDelay(1.);
-    for (const rosbag::MessageInstance& msg : view) {
-      if (!::ros::ok()) {
+  for (size_t bag_index = 0; bag_index < bag_filenames.size();) {
+    const Event message = queue.Pop();
+    switch (message.kind) {
+      case Event::Kind::kNewBag:
+        begin_time = message.new_bag.begin_time;
+        duration_in_seconds = message.new_bag.duration_in_seconds;
+        trajectory_id =
+            node.AddOfflineTrajectory(expected_sensor_ids, trajectory_options);
+        break;
+
+      case Event::Kind::kBagEnd:
+        node.FinishTrajectory(trajectory_id);
+        ++bag_index;
+        break;
+
+      case Event::Kind::kRosMessage: {
+        if (FLAGS_use_bag_transforms &&
+            message.message.kind == Event::RosMessage::Kind::kTf) {
+          tf_publisher.publish(*message.message.tf);
+          for (const auto& transform : message.message.tf->transforms) {
+            try {
+              tf_buffer.setTransform(transform, "unused_authority",
+                                     message.message.topic == kTfStaticTopic);
+            } catch (const tf2::TransformException& ex) {
+              LOG(WARNING) << ex.what();
+            }
+          }
+          // NOCOM(#hrapp): ugly.
+          break;  // Alread
+        }
+
+        while (!delayed_messages.empty() &&
+               delayed_messages.front().serialization_time <
+                   message.message.serialization_time - kDelay) {
+          const Event::RosMessage& delayed_msg = delayed_messages.front();
+          CHECK_NE(trajectory_id, -1)
+              << "Did not see a kNewBag message before first ROS message.";
+          switch (delayed_msg.kind) {
+            case Event::RosMessage::Kind::kLaserScan:
+              node.HandleLaserScanMessage(trajectory_id, delayed_msg.topic,
+                                          delayed_msg.laser_scan);
+              break;
+
+            case Event::RosMessage::Kind::kMultiEchoLaserScan:
+              node.HandleMultiEchoLaserScanMessage(
+                  trajectory_id, delayed_msg.topic,
+                  delayed_msg.multi_echo_laser_scan);
+              break;
+
+            case Event::RosMessage::Kind::kPointCloud2:
+              node.HandlePointCloud2Message(trajectory_id, delayed_msg.topic,
+                                            delayed_msg.point_cloud2);
+              break;
+
+            case Event::RosMessage::Kind::kImu:
+              node.HandleImuMessage(trajectory_id, delayed_msg.topic,
+                                    delayed_msg.imu);
+              break;
+
+            case Event::RosMessage::Kind::kOdometry:
+              node.HandleOdometryMessage(trajectory_id, delayed_msg.topic,
+                                         delayed_msg.odometry);
+              break;
+
+            case Event::RosMessage::Kind::kTf:
+              LOG(FATAL) << "Should already been handled.";
+              break;
+          }
+
+          clock.clock = delayed_msg.serialization_time;
+          clock_publisher.publish(clock);
+
+          LOG_EVERY_N(INFO, 100000)
+              << "Processed "
+              << (delayed_msg.serialization_time - begin_time).toSec() << " of "
+              << duration_in_seconds << " bag time seconds...";
+          delayed_messages.pop_front();
+        }
+        delayed_messages.push_back(message.message);
         break;
       }
-
-      if (FLAGS_use_bag_transforms && msg.isType<tf2_msgs::TFMessage>()) {
-        auto tf_message = msg.instantiate<tf2_msgs::TFMessage>();
-        tf_publisher.publish(tf_message);
-
-        for (const auto& transform : tf_message->transforms) {
-          try {
-            tf_buffer.setTransform(transform, "unused_authority",
-                                   msg.getTopic() == kTfStaticTopic);
-          } catch (const tf2::TransformException& ex) {
-            LOG(WARNING) << ex.what();
-          }
-        }
-      }
-
-      while (!delayed_messages.empty() &&
-             delayed_messages.front().getTime() < msg.getTime() - kDelay) {
-        const rosbag::MessageInstance& delayed_msg = delayed_messages.front();
-        const std::string topic = node.node_handle()->resolveName(
-            delayed_msg.getTopic(), false /* resolve */);
-        if (delayed_msg.isType<sensor_msgs::LaserScan>()) {
-          node.HandleLaserScanMessage(
-              trajectory_id, topic,
-              delayed_msg.instantiate<sensor_msgs::LaserScan>());
-        }
-        if (delayed_msg.isType<sensor_msgs::MultiEchoLaserScan>()) {
-          node.HandleMultiEchoLaserScanMessage(
-              trajectory_id, topic,
-              delayed_msg.instantiate<sensor_msgs::MultiEchoLaserScan>());
-        }
-        if (delayed_msg.isType<sensor_msgs::PointCloud2>()) {
-          node.HandlePointCloud2Message(
-              trajectory_id, topic,
-              delayed_msg.instantiate<sensor_msgs::PointCloud2>());
-        }
-        if (delayed_msg.isType<sensor_msgs::Imu>()) {
-          node.HandleImuMessage(trajectory_id, topic,
-                                delayed_msg.instantiate<sensor_msgs::Imu>());
-        }
-        if (delayed_msg.isType<nav_msgs::Odometry>()) {
-          node.HandleOdometryMessage(
-              trajectory_id, topic,
-              delayed_msg.instantiate<nav_msgs::Odometry>());
-        }
-        clock.clock = delayed_msg.getTime();
-        clock_publisher.publish(clock);
-
-        LOG_EVERY_N(INFO, 100000)
-            << "Processed " << (delayed_msg.getTime() - begin_time).toSec()
-            << " of " << duration_in_seconds << " bag time seconds...";
-
-        delayed_messages.pop_front();
-      }
-
-      const std::string topic =
-          node.node_handle()->resolveName(msg.getTopic(), false /* resolve */);
-      if (expected_sensor_ids.count(topic) == 0) {
-        continue;
-      }
-      delayed_messages.push_back(msg);
     }
-
-    bag.close();
-    node.FinishTrajectory(trajectory_id);
   }
+
+  rosbag_io_thread.join();
 
   // Ensure the clock is republished after the bag has been finished, during the
   // final optimization, serialization, and optional indefinite spinning at the