Timings using `/usr/bin/time roslaunch cartographer_ros offline_backpack_3d.launch bag_filenames:=$HOME/Downloads/b3-2015-12-10-13-55-20.bag` after removing `occupancy_grid_node`, rViz, and setting `required="True"` on the node.

master:
Elapsed wall clock time: 232.837 s (100%)
Elapsed CPU time: 386.774 s (100%)
Peak memory usage: 1515016 KiB (100%)
372.32user (100%) 68.22system (100%) 4:04.33 (100%) elapsed 180%CPU k

this branch:
Elapsed wall clock time: 211.506 s (90.8%)
Elapsed CPU time: 386.884 s (100%)
Peak memory usage: 1554836 KiB (102.6%)
375.02user (100.8%) 62.09system (100%) 3:43.10elapsed (94.39%) 195%CPU