@SirVer 2017-11-06
@cschuet @wohe Is getting 10% of your lifetime back worth the increase in code and complexity? I am quite torn - I was also hoping for higher gains.