@SirVer 2017-11-08
@wohe I tried the lambda approach in https://github.com/SirVer/cartographer_ros/tree/01_multithread_use_lambdas, unfortunately it takes ~150% of current master, so is much slower.