@SirVer 2017-11-10
After some more experiments I am not convinced that this will have the impact that I was hoping for. I think the rfc 0, i.e. moving towards a gRPC based solution and wrapping this in ROS will have similar results (i.e. prereading in the ROS thread will happen) and should be cleaner. Abandoning this for the moment.